최신정보 게시판 오류 수정

+++ src/main/java/com/teamproject/devTalks/service/implement/user/UserServiceImplement.java
@@ -249,7 +249,7 @@ public class UserServiceImplement implements UserService {
 
             List<UserHashtagEntity> currentUserHashTagEntities = userHashtagRepository.findAllByUserNumber(userNumber);
 
-            userHashtagRepository.deleteAll(currentUserHashTagEntities);
+            if(currentUserHashTagEntities != null )userHashtagRepository.deleteAll(currentUserHashTagEntities);
 
             for (String hashtag : hashtagList) {
